[FEATURE]#40 : 기존 MessageScreen에서, 홈화면 분리
